Erase MAPQ for alignments set to unmapped (#50)
As #33 and #35 pointed out, there are downstream errors (eg `picard
MarkDuplicates`) resulting from our initial design where the MAPQ of
unliftable, deferred, and suppressed reads are kept. To avoid such
errors, I changed the default behavior of leviosam2 so that the
resulting MAPQ is 0 if a read is assigned to unmapped during lift-over.

To keep MAPQs, users may activate the `--keep_mapq` flag, which will
keep the original MAPQ values.