CRAM support in the python workflow (#66)
This PR supports writing to the CRAM format in the Python workflow using
the `-O cram` option, suggested by @insilicool in
https://github.com/milkschen/leviosam2/issues/64.

I didn't change the core `leviosam2` software, but added a command after
the main workflow to convert the final BAM file to the CRAM format. It
would be interesting to see if it's possible to write a real CRAM file
in the leviosam2 program, but for now I think we can use this solution.