Calculates the ref length of mate read using "MC:Z" when available (#67)
Previously, we inferred the mate reference length using the query length
of a read. This can result in inconsistent lift-over results between a
pair of reads. For example, if the soft clipped region of a paired-end
read overlaps with a chain gap, leviosam2 used to ignore the gap for the
main segment but assign the mate to be unliftable.

The `MC:Z` tag describes the CIGAR of the mate read in the paired-end
mode. This PR uses the tag (when available) to calculate the mate
starting and ending positions (wrt the reference) to increase the
consistency of paired-end lift-over.

This feature is not compatible with htslib v1.11, which is pretty old. I
also terminate the support of that version in this PR.