Errors out when the chain map validator finds an issue (#92)
The leviosam2 algorithm does not support the case when two or more
intervals in the source genome overlap. Previously an error message was
displayed, but the program could still proceed. In this PR, we update
the program behavior to error out when validation fails.

The below shows the functional change. Other changed lines are tests and
docs.
```
    if (!validate_intervals()) {
        std::cerr << "[E::chain::build] Interval validation failed\n";
        exit(1);
    }
```